requirements.txt
pandas
Pillow
natsort
matplotlib
scikit-image
tqdm
numpy
shapely